requirements.txt
backports.functools-lru-cache==2.0.0
blinker==1.7.0
click==8.1.7
DBUtils==3.0.3
et-xmlfile==1.1.0
Flask==3.0.0
indic-transliteration==2.3.54
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysqlclient==2.2.1
openpyxl==3.1.2
python-dotenv==1.0.0
regex==2023.10.3
roman==4.1
toml==0.10.2
typer==0.9.0
typing_extensions==4.9.0
Werkzeug==3.0.1
